refactor: move class after package creation
