+++ src/main/java/RacingCarApplication.java
@@ -1,4 +1,4 @@
-import controller.RacingGame;
+import racingcar.controller.RacingGame;
 
 public class RacingCarApplication {
     public static void main(String[] args) {
